+++ src/ZoneServer/Buffs/Handlers/Scout/FreeStep_Buff.cs
@@ -1,7 +1,7 @@
 ﻿using Melia.Shared.Game.Const;
 using Melia.Zone.Buffs.Base;
 
-namespace Melia.Zone.Buffs.Handlers
+namespace Melia.Zone.Buffs.Handlers.Scout
 {
 	/// <summary>
 	/// Handle for the Free Step Buff, which increases the target's
